social_media/requirements.txt
Django
django-birthday
Babel
oauthlib
Pillow
psycopg2
PyJWT
python-dotenv
